src/api/requirements.txt
fastapi
fastapi[standard]
requests
pandas

azure-monitor-opentelemetry
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai-v2
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp-proto-grpc

jsonlines
python-dotenv==1.1.0
gunicorn==23.0.0
jupyter

prompty[azure]==0.1.50
azure-search-documents==11.4.0
azure-identity==1.21.0
azure-keyvault-secrets
azure-ai-evaluation
azure-ai-inference[opentelemetry]
azure-ai-projects==1.0.0b9
azure-ai-contentsafety
azure-ai-ml
azureml-mlflow

aiohttp==3.10.11
nbconvert
validators
tenacity
marshmallow==3.23.2
mcp==1.6.0
semantic-kernel==1.28.1

